[refactor] ThreadItem component common 폴더로 이동
